Documentation Change - Username Policies - Option2
Adding two different requests with suggestions for improved versions of the documentation in place. We don't currently mention the Username policies (limitations for allowed characters and string lengths) in this documentation. In my view, it would be a useful addition to the current status. This SECOND one adds a bullet point to the list instead of a full paragraph mentioning the guest user UPN generation behavior.